chore: move css prop jsx element creation into utils
